requirements.txt
botbuilder-core>=4.15.0
botbuilder-schema>=4.15.0
botbuilder-integration-aiohttp>=4.15.0
aiohttp>=3.8.0
python-dotenv==1.1.0
azure-identity==1.23.0
openai==1.78.1
pydantic==2.11.4
pydantic-settings==2.2.1
rich==14.0.0
